+++ lib/rspec/rails/example/helper_example_group.rb
@@ -36,6 +36,10 @@ module RSpec::Rails
       include Webrat::Matchers
     end
 
+    capybara do
+      include Capybara::RSpecMatchers if defined?(Capybara::RSpecMatchers)
+    end
+
     module ClassMethods
       def determine_default_helper_class(ignore)
         describes